Move Java projects out of `java` package
