+++ modules/models/java-project/src/main/kotlin/org/cafejojo/schaapi/models/project/JavaMavenProject.kt
@@ -1,6 +1,5 @@
-package org.cafejojo.schaapi.models.project.java
+package org.cafejojo.schaapi.models.project
 
-import org.cafejojo.schaapi.models.project.MavenProject
 import java.io.File
 
 /**
