+++ modules/pipeline/java-maven-project-compiler/src/main/kotlin/org/cafejojo/schaapi/pipeline/projectcompiler/javamaven/ProjectCompiler.kt
@@ -3,7 +3,7 @@ package org.cafejojo.schaapi.pipeline.projectcompiler.javamaven
 import org.apache.maven.shared.invoker.DefaultInvocationRequest
 import org.apache.maven.shared.invoker.DefaultInvoker
 import org.cafejojo.schaapi.models.Project
-import org.cafejojo.schaapi.models.project.java.JavaMavenProject
+import org.cafejojo.schaapi.models.project.JavaMavenProject
 import org.cafejojo.schaapi.pipeline.ProjectCompiler
 import java.io.File
