+++ modules/pipeline/jimple-evosuite-test-generator/src/main/kotlin/org/cafejojo/schaapi/pipeline/testgenerator/jimpleevosuite/TestGenerator.kt
@@ -1,6 +1,6 @@
 package org.cafejojo.schaapi.pipeline.testgenerator.jimpleevosuite
 
-import org.cafejojo.schaapi.models.project.java.JavaProject
+import org.cafejojo.schaapi.models.project.JavaProject
 import org.cafejojo.schaapi.pipeline.Pattern
 import org.cafejojo.schaapi.pipeline.TestGenerator
 import java.io.File
