+++ modules/pipeline/jimple-library-usage-graph-generator/src/main/kotlin/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/LibraryUsageGraphGenerator.kt
@@ -4,7 +4,7 @@ import org.cafejojo.schaapi.models.DfsIterator
 import org.cafejojo.schaapi.models.Node
 import org.cafejojo.schaapi.models.Project
 import org.cafejojo.schaapi.models.libraryusagegraph.jimple.JimpleNode
-import org.cafejojo.schaapi.models.project.java.JavaProject
+import org.cafejojo.schaapi.models.project.JavaProject
 import org.cafejojo.schaapi.pipeline.LibraryUsageGraphGenerator
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.filters.BranchStatementFilter
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.filters.StatementFilter
