+++ modules/pipeline/jimple-library-usage-graph-generator/src/main/kotlin/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/filters/BranchStatementFilter.kt
@@ -1,6 +1,6 @@
 package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.filters
 
-import org.cafejojo.schaapi.models.project.java.JavaProject
+import org.cafejojo.schaapi.models.project.JavaProject
 import soot.Body
 import soot.Unit
 import soot.Value
